[ADD] Add new hook point for ECCUBE version >= 3.0.9
